Rename Theme.interface to types in theme folder
