+++ TabbedTSApp/src/screens/MasterDetail/MasterDetailScreen.style.ts
@@ -1,5 +1,5 @@
 import {StyleSheet} from 'react-native';
-import {Theme} from '../../themes/Theme.interface';
+import {Theme} from '../../themes/types';
 
 export const getStyles = (theme: Theme) =>
   StyleSheet.create({
